+++ pyincore/analyses/commercialbuildingrecovery/commercialbuildingrecovery.py
@@ -67,11 +67,12 @@ class CommercialBuildingRecovery(BaseAnalysis):
         sample_damage_states = self.get_input_dataset("sample_damage_states").get_dataframe_from_csv(low_memory=False)
         mcs_failure = self.get_input_dataset("mcs_failure").get_dataframe_from_csv(low_memory=False)
         redi_delay_factors = self.get_input_dataset("delay_factors").get_dataframe_from_csv(low_memory=False)
+        building_dmg = self.get_input_dataset("building_dmg").get_dataframe_from_csv(low_memory=False)
 
         # Returns dataframe
         total_delay, recovery, time_stepping_recovery = self.commercial_recovery(buildings, sample_damage_states,
                                                                                  mcs_failure, redi_delay_factors,
-                                                                                 num_samples)
+                                                                                 building_dmg, num_samples)
         self.set_result_csv_data("total_delay", total_delay, result_name + "_delay", "dataframe")
         self.set_result_csv_data("recovery", recovery, result_name + "_recovery", "dataframe")
         self.set_result_csv_data("time_stepping_recovery", time_stepping_recovery,
@@ -79,7 +80,7 @@ class CommercialBuildingRecovery(BaseAnalysis):
 
         return True
 
-    def commercial_recovery(self, buildings, sample_damage_states, mcs_failure, redi_delay_factors, num_samples):
+    def commercial_recovery(self, buildings, sample_damage_states, mcs_failure, redi_delay_factors, building_dmg, num_samples):
         """
         Calculates commercial building recovery for buildings
 
@@ -88,6 +89,7 @@ class CommercialBuildingRecovery(BaseAnalysis):
             sample_damage_states (pd.DataFrame): Sample damage states
             redi_delay_factors (pd.DataFrame): Delay factors based on REDi framework
             mcs_failure (pd.DataFrame): Building inventory failure probabilities
+            building_dmg (pd.DataFrame): Building damage states
             num_samples (int): number of sample scenarios to use
 
         Returns:
@@ -97,7 +99,7 @@ class CommercialBuildingRecovery(BaseAnalysis):
 
         start_total_delay = time.process_time()
         total_delay = CommercialBuildingRecovery.total_delay(buildings, sample_damage_states, mcs_failure,
-                                                             redi_delay_factors, num_samples)
+                                                             redi_delay_factors, building_dmg, num_samples)
         end_total_delay = time.process_time()
         print("Finished executing total_delay() in " + str(end_total_delay - start_total_delay) + " secs")
 
@@ -116,7 +118,7 @@ class CommercialBuildingRecovery(BaseAnalysis):
         return total_delay, recovery, time_stepping_recovery
 
     @staticmethod
-    def total_delay(buildings, sample_damage_states, mcs_failure, redi_delay_factors, num_samples):
+    def total_delay(buildings, sample_damage_states, mcs_failure, redi_delay_factors, damage, num_samples):
         """ Calculates total delay by combining financial delay and other factors from REDi framework
 
         Args:
@@ -125,6 +127,7 @@ class CommercialBuildingRecovery(BaseAnalysis):
             mcs_failure (pd.DataFrame): Building inventory failure probabilities
             redi_delay_factors (pd.DataFrame): Delay impeding factors such as post-disaster inspection, insurance claim,
                 financing, and government permit based on building's damage state.
+            damage (pd.DataFrame): Damage states for building structural damage
             num_samples (int): number of sample scenarios to use
 
         Returns:
@@ -132,7 +135,7 @@ class CommercialBuildingRecovery(BaseAnalysis):
         """
 
         # Obtain the commercial buildings in damage
-        damage = mcs_failure[mcs_failure['haz_expose'] == 'yes']
+        damage = mcs_failure[damage['haz_expose'] == 'yes']
         commercial = []
         commercial_archetypes = [6, 7, 8, 15, 16, 18, 19]
         for i, b in enumerate(buildings):
@@ -398,6 +401,12 @@ class CommercialBuildingRecovery(BaseAnalysis):
                     'description': 'Delay impeding factors such as post-disaster inspection, insurance claim, '
                                    'and government permit based on building\'s damage state. Provided by REDi framework',
                     'type': ['incore:buildingRecoveryFactors']
+                },
+                {
+                    'id': 'building_dmg',
+                    'required': True,
+                    'description': 'damage result that has damage intervals',
+                    'type': ['ergo:buildingDamageVer6']
                 }
             ],
             'output_datasets': [